backend/requirements.txt
django == 2.2.0
psycopg2-binary == 2.8.4
djangorestframework == 3.10.3
celery==4.4.0
